+++ OSPaymentsLib/Protocols/OSPMTGatewayDelegate.swift
@@ -9,8 +9,9 @@ protocol OSPMTGatewayDelegate: AnyObject {
     /// - Parameters:
     ///   - payment: Apple Pay's payment request result.
     ///   - details: Payment details to trigger processing.
+    ///   - accessToken: Authorisation token related with a full payment type.
     ///   - completion: Payment process result. If returns the process result in case of success or an error otherwise.
-    func process(_ payment: PKPayment, with details: OSPMTDetailsModel, _ completion: @escaping (Result<OSPMTServiceProviderInfoModel, OSPMTError>) -> Void)
+    func process(_ payment: PKPayment, with details: OSPMTDetailsModel, and accessToken: String, _ completion: @escaping (Result<OSPMTServiceProviderInfoModel, OSPMTError>) -> Void)
 }
 
 extension OSPMTGatewayDelegate {
@@ -18,11 +19,14 @@ extension OSPMTGatewayDelegate {
     /// Triggers the backend url request.
     /// - Parameters:
     ///   - requestParameters: Model containing the request body to trigger
+    ///   - accessToken: Authorisation token related with a full payment type.
     ///   - completion: Payment process result. If returns the process result in case of success or an error otherwise.
-    func processURLRequest(_ requestParameters: OSPMTRequestParametersModel, _ completion: @escaping (Result<OSPMTServiceProviderInfoModel, OSPMTError>) -> Void) {
+    func processURLRequest(_ requestParameters: OSPMTRequestParametersModel, and accessToken: String, _ completion: @escaping (Result<OSPMTServiceProviderInfoModel, OSPMTError>) -> Void) {
         self.urlRequest.httpMethod = "POST"
         self.urlRequest.setValue("application/json", forHTTPHeaderField: "Content-Type")
         self.urlRequest.setValue("application/json", forHTTPHeaderField: "Accept")
+        self.urlRequest.setValue(accessToken, forHTTPHeaderField: "Payments-Token")
+        
         self.urlRequest.httpBody = try? JSONEncoder().encode(requestParameters)
         let task = self.urlSession.dataTask(with: self.urlRequest) { data, response, error in
             guard let response = response as? HTTPURLResponse,